Update switching logic to use 4 bytes
Added enums for header types. Once the network headers are stripped
from the PCAP file there are still leftover bytes that need to be
stripped out that always start with 0xBB0BBB0B. I created a header to
catch these junk bytes. For the ASCII logs I am using the first four
bytes of the message for #RXCOMMANDS it would be #RXC

enums:
  novatel_header_types:
    0xBB0BBB0B: bbb_junk
    0x1c1244aa: novatel_binary_legacy_header
    0x5647ccaa: novatel_binary_g3_header
    0x43585223: novatel_ascii_rxcommands