Finished first version
Ran through a whole G-III file from HNL-C and a GUST file from AP1. There is a
lot of *_stub* code in this, meaning I made all the logic to handle any message,
but I didn't fully parse every message, but only calculated the message length
and put the contents in a byte array.